Function hacspec_halo2::step_5

pub(crate) fn step_5(h: Seq<Fp>, n: u128) -> Seq<Seq<Fp>>

Step 5 split polynomial of degree n_g(n-1)-n up into n_(g-2) polynomials of degree at most n-1

The polynomials(represented by vectors) are stored in a vectore. This way the index in the outer vector can act as the i when reproducing the original poly: h(X) = SUM from i=0 to n_(g-1) [xˆ(ni)h_i(x)] Where n is a parameter of the prooving system, and h_i is the ith part of the original poly.

Arguments

h - Polynomial to be split

n - defines length of new polynomials (global variable for prooving system)