Type Definition hacspec_halo2::Term

pub(crate) type Term = (Fp, Seq<u32>);

A term in a multivariate polynomial

Tuple entries

0: - The first entry represents the coefficient for the term.

1: - The second entry, a sequence of u32s, represent the powers for each variable, s.t. the i’th variabe is raised to the power of the i’th entry in the sequence.